fix: sonarqube duplicated code
